# lxdweb/docs/swagger.yaml
basePath: /
definitions:
  models.CreateIPv6Request:
    properties:
      container_hostname:
        type: string
      description:
        type: string
      node_id:
        type: integer
    required:
    - container_hostname
    - node_id
    type: object
...
host: localhost:3000
info:
  contact:
    email: xkatld@example.com
    name: xkatld
    url: https://github.com/xkatld/zjmf-lxd-server
  description: 基于 Session 的认证
  license:
    name: MIT
    url: https://github.com/xkatld/zjmf-lxd-server/blob/main/LICENSE
  termsOfService: https://github.com/xkatld/zjmf-lxd-server
  title: LXD Web 管理平台
  version: 1.0.3
paths:
...
  /api/ipv6:
    get:
      description: 查询所有IPv6绑定信息，支持按节点过滤
      parameters:
      - description: 节点ID
        in: query
        name: node_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回IPv6绑定列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 查询失败
          schema:
            additionalProperties: true
            type: object
      summary: 获取IPv6绑定列表
      tags:
      - IPv6管理
    post:
      consumes:
      - application/json
      description: 为指定容器创建新的IPv6地址绑定
      parameters:
      - description: IPv6绑定参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateIPv6Request'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 节点不存在
          schema:
            additionalProperties: true
            type: object
      summary: 创建IPv6绑定
      tags:
      - IPv6管理
  /api/ipv6-sync/all:
    post:
      description: 启动所有活跃节点的IPv6绑定同步任务
      produces:
      - application/json
      responses:
        "200":
          description: 同步任务已启动
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未登录
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 查询失败
          schema:
            additionalProperties: true
            type: object
      summary: 同步所有节点的IPv6绑定
      tags:
      - IPv6管理
  /api/ipv6-sync/status:
    get:
      description: 查询指定节点或所有节点的IPv6绑定同步状态
      parameters:
      - description: 节点ID
        in: query
        name: node_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回同步状态
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未登录
          schema:
            additionalProperties: true
            type: object
      summary: 获取IPv6同步状态
      tags:
      - IPv6管理
  /api/ipv6/{id}:
    delete:
      description: 删除指定的IPv6地址绑定并清理数据库
      parameters:
      - description: 绑定ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 绑定或节点不存在
          schema:
            additionalProperties: true
            type: object
      summary: 删除IPv6绑定
      tags:
      - IPv6管理
  /api/ipv6/cache:
    get:
      description: 直接从本地缓存数据库读取IPv6绑定信息
      parameters:
      - description: 节点ID
        in: query
        name: node_id
        type: string
      - description: 容器名称
        in: query
        name: hostname
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回缓存数据
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 查询失败
          schema:
            additionalProperties: true
            type: object
      summary: 从缓存获取IPv6绑定
      tags:
      - IPv6管理
  /api/ipv6/sync:
    post:
      description: 从lxdapi同步指定节点的IPv6绑定信息
      parameters:
      - description: 节点ID
        in: query
        name: node_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 同步任务已启动
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 缺少参数
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 节点不存在
          schema:
            additionalProperties: true
            type: object
      summary: 同步IPv6绑定
      tags:
      - IPv6管理
  /api/ipv6/tasks:
    get:
      description: 查询最近50条IPv6同步任务记录
      parameters:
      - description: 节点ID
        in: query
        name: node_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回任务列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 查询失败
          schema:
            additionalProperties: true
            type: object
      summary: 获取IPv6同步任务列表
      tags:
      - IPv6管理
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"